requirements.txt
Django==5.1.3
pillow==11.0.0
psycopg2==2.9.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0